# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wiles and Rues API",
    "version": "1.0.0",
    "description": "REST API for the Wiles and Rues e-commerce store — authentication, products, reviews, orders (Stripe checkout + webhooks), blogs, policies, feedback, contact and admin stats.\n\n**Authentication** — log in with `POST /api/auth/login`; the server sets an httpOnly cookie named `token`. Use the Authorize button (cookie) or send `Authorization: Bearer <token>` on protected routes. Admin-only routes return `403` for non-admin users.",
    "contact": {
      "name": "Wiles and Rues"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development"
    },
    {
      "url": "https://e-commerce-project-hzpl.onrender.com/api",
      "description": "Production (Render)"
    }
  ],
...
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all users (admin)",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page, max 100 (default 10)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Partial email search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products (paginated, filterable)",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Category filter (use `all` to clear)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "color",
            "in": "query",
            "description": "Color filter (use `all` to clear)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Case-insensitive name search",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page, max 100 (default 10)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated product list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/orders/webhook": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Stripe webhook",
        "description": "Receives `checkout.session.completed` (records paid orders + decrements stock) and `payment_intent.payment_failed` (marks orders canceled). Requires a valid `Stripe-Signature` header — the body must be sent raw.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Raw Stripe event payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List all orders (admin, paginated + filterable)",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page, max 100 (default 10)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status: pending, processing, shipped, delivered, canceled",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Partial email search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "tags": [
          "Contact"
        ],
        "summary": "Send a contact form message",
        "description": "Emails the store inbox via Resend. Returns 500 when email is not configured.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/blogs": {
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "List published blogs (paginated)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page, max 100 (default 12)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "includeDrafts",
            "in": "query",
            "description": "Include unpublished posts (admin only)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated blog list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/blogs/admin": {
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "List all blogs incl. drafts (admin)",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page, max 100 (default 12)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated blog list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/policy": {
      "get": {
        "tags": [
          "Policy"
        ],
        "summary": "Get the store policy",
        "responses": {
          "200": {
            "description": "Policy document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Policy"
        ],
        "summary": "Update the store policy (admin)",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Policy updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyWriteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/feedback": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "summary": "List all feedback (admin, filterable)",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "new, in_progress, resolved or rejected",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "bug or feature",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feedback"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Feedback"
        ],
        "summary": "Submit feedback (bug or feature)",
        "description": "Authentication is optional — signed-in users get their identity attached to the report. Guests submit anonymously.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackWriteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/upload-image": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Upload an image to Cloudinary (admin)",
        "description": "Body is a JSON object with a base64 data-URI `image` field.",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadImageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT session cookie set by login/register."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Alternative: send the JWT in the Authorization header."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed or malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, invalid or expired session",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed (admin-only, or not your own resource)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "description": "Sometimes present with technical details"
          }
        }
      },
      "ObjectId": {
        "type": "string",
        "description": "MongoDB ObjectId (24 hex chars)",
        "pattern": "^[a-fA-F0-9]{24}$"
      },
...
      "UserListResponse": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "$ref": "#/components/schemas/ObjectId"
                },
                "email": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          },
          "totalUsers": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "oldPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer",
            "nullable": true,
            "description": "Absent/null means unlimited stock"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "color": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "author": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "totalProducts": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          }
        }
      },
...
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "received": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "orderId": {
            "type": "string",
            "description": "Stripe payment intent id"
          },
          "products": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "$ref": "#/components/schemas/ObjectId"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "amount": {
            "type": "number",
            "description": "Order total in dollars"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "shipped",
              "delivered",
              "canceled"
            ]
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "time": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/ShippingAddress"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderListResponse": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "totalOrders": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          }
        }
      },
...
      "ContactRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "subject",
          "message"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "subject": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "message": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000
          }
        }
      },
      "ContactResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Resend email id"
          }
        }
      },
...
      "BlogListResponse": {
        "type": "object",
        "properties": {
          "blogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Blog"
            }
          },
          "totalBlogs": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          }
        }
      },
...
      "PolicyResponse": {
        "type": "object",
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/Policy"
          }
        }
      },
      "UpdatePolicyRequest": {
        "type": "object",
        "required": [
          "title",
          "introduction",
          "sections"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 160
          },
          "introduction": {
            "type": "string",
            "minLength": 20,
            "maxLength": 2000
          },
          "sections": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/PolicySection"
            }
          }
        }
      },
      "PolicyWriteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          }
        }
      },
      "FeedbackUser": {
        "type": "object",
        "properties": {
          "userId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ],
            "description": "Null for anonymous (guest) feedback"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "Feedback": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "type": {
            "type": "string",
            "enum": [
              "bug",
              "feature"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_progress",
              "resolved",
              "rejected"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/FeedbackUser"
          },
          "pageUrl": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateFeedbackRequest": {
        "type": "object",
        "required": [
          "type",
          "title",
          "description"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "bug",
              "feature"
            ]
          },
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 3000
          },
          "pageUrl": {
            "type": "string",
            "maxLength": 1000
          }
        }
      },
...
      "FeedbackWriteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "feedback": {
            "$ref": "#/components/schemas/Feedback"
          }
        }
      },
...
      "UploadImageRequest": {
        "type": "object",
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "description": "Base64 data-URI, e.g. data:image/png;base64,..."
          }
        }
      },
      "UploadImageResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      }
    }
  }
}